revert boxCast refactor (while it respects value classes, it is not specifically for value classes, so it makes more sense to be in a general util class)
