+++ modules/mockk-core/src/commonMain/kotlin/io/mockk/platform/ValueClassSupport.kt
@@ -19,14 +19,4 @@ expect object ValueClassSupport {
      */
     val KClass<*>.boxedClass: KClass<*>
 
-    /**
-     * Normally this simply casts [arg] to `T`
-     *
-     * However, if `T` is a `value class` (of type [cls]) this will construct a new instance of the
-     * value class, and set [arg] as the value.
-     */
-      fun <T : Any> boxCast(
-        cls: KClass<*>,
-        arg: Any,
-    ): T
 }
