+++ modules/mockk-core/src/jvmMain/kotlin/io/mockk/platform/ValueClassSupport.kt
@@ -65,16 +65,4 @@ actual object ValueClassSupport {
             false
         }
 
-    @Suppress("UNCHECKED_CAST")
-    actual fun <T : Any> boxCast(
-        cls: KClass<*>,
-        arg: Any,
-    ): T {
-        return if (cls.isValue) {
-            val constructor = cls.primaryConstructor!!.apply { isAccessible = true }
-            constructor.call(arg) as T
-        } else {
-            arg as T
-        }
-    }
 }
