+++ modules/mockk-dsl/src/commonMain/kotlin/io/mockk/InternalPlatformDsl.kt
@@ -1,6 +1,7 @@
 package io.mockk
 
 import kotlin.coroutines.Continuation
+import kotlin.reflect.KClass
 
 expect object InternalPlatformDsl {
     fun identityHashCode(obj: Any): Int
@@ -35,6 +36,17 @@ expect object InternalPlatformDsl {
     fun counter(): InternalCounter
 
     fun <T> coroutineCall(lambda: suspend () -> T): CoroutineCall<T>
+
+    /**
+     * Normally this simply casts [arg] to `T`
+     *
+     * However, if `T` is a `value class` (of type [cls]) this will construct a new instance of the
+     * value class, and set [arg] as the value.
+     */
+    fun <T : Any> boxCast(
+        cls: KClass<*>,
+        arg: Any,
+    ): T
 }
 
 interface CoroutineCall<T> {
