+++ modules/mockk-dsl/src/jvmMain/kotlin/io/mockk/InternalPlatformDsl.kt
@@ -1,6 +1,5 @@
 package io.mockk
 
-import io.mockk.platform.ValueClassSupport.boxedClass
 import kotlinx.coroutines.runBlocking
 import java.lang.reflect.AccessibleObject
 import java.lang.reflect.InvocationTargetException
@@ -215,6 +214,18 @@ actual object InternalPlatformDsl {
 
     actual fun <T> coroutineCall(lambda: suspend () -> T): CoroutineCall<T> = JvmCoroutineCall<T>(lambda)
 
+    @Suppress("UNCHECKED_CAST")
+    actual fun <T : Any> boxCast(
+        cls: KClass<*>,
+        arg: Any,
+    ): T {
+        return if (cls.isValue) {
+            val constructor = cls.primaryConstructor!!.apply { isAccessible = true }
+            constructor.call(arg) as T
+        } else {
+            arg as T
+        }
+    }
 }
 
 class JvmCoroutineCall<T>(private val lambda: suspend () -> T) : CoroutineCall<T> {
